Updated to the current standard.
Adjusted the names of the views, panel, and view panes to the current standard.

Updated the taxonomy alias to be the current standard.

Split the views in to list, node, admin, and sort.

Added category for the panel panes to list, node, admin, and sort.

Added the imce profile install.

Signed-off-by: Staris <jaison@worxco.com>