Updated to the feature on the newest platform.
Added taxonomy term pathauto changes so that paths with other features won't conflict.

Added css for the term and to vertically align the table cells to the top.

Changed the display name for Owner to Contact.

Added Admin configurations for the Menu and Title.

Moved the pathauto for the term to the install file so the path can be changed without overriding the feature.

Signed-off-by: Staris <jaison@worxco.com>